pyproject.toml
[tool.poetry]
name = "cleo"
version = "0.3.0"
description = "Cleo"
authors = ["Milton Abrunhosa <miltonmla@decea.mil.br>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
streamlit = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pylint = "^2.13.9"
mypy = "^0.950"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
pip-audit = "^2.2.1"
safety = "^1.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
